requirements.txt
amqp==5.0.2
asgiref==3.2.10
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.5
celery-progress==0.0.14
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-admin-autocomplete-filter==0.6.1
django-admin-autocomplete-list-filter==0.1.3
django-admin-list-filter-dropdown==1.0.3
django-advanced-filters==1.2.0
django-bootstrap4==2.3.1
django-braces==1.13.0
django-cors-headers==3.5.0
django-extensions==3.1.0
django-filter==2.4.0
django-group-by==0.3.1
django-import-export==2.4.0
django-model-utils==4.0.0
django-rest-multiple-models==2.1.3
djangorestframework==3.12.1
et-xmlfile==1.0.1
idna==2.10
install==1.3.4
jdcal==1.4.1
kombu==5.0.2
Markdown==3.3.3
MarkupPy==1.14
numpy==1.19.2
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.4
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
django-ckeditor==6.0.0
django-summernote==0.8.11.6
django-rest-auth==0.9.5
django-allauth==0.44.0
